+++ build-scripts/webpack.config.js
@@ -40,6 +40,7 @@ module.exports = env => {
             ],
         },
         devServer: {
+            writeToDisk: true,
             useLocalIp: true,
             host: "0.0.0.0",
             historyApiFallback: {